Improved player animations. Added hotbar
Fixed issue where, during the attack animation, the hitbox of the weapon moved in a way that made it hard to hit enemies. Hitbox therefore no longer lines up with the position of the weapon during the animation but is at least effective at hitting enemies. Added hotbar that allows you to switch between gameobjects. Added Axe.