Added dropped items and, optimized and changed inventory system
Dropped items can be picked up when the player is near it and presses F. When the inventory was activated the hotbar would automatically close (intentionally) but it has now been changed so that the inventory won't change whether or not the hotbar is active. Also optimized and shortened down the code for moving items.