Added enemy AI searching (for player) system, somewhat fixed enemy animation problem
The enemy walking animation looks really bad when apply motion is turned off. So, the enemy have had apply motion turned on but it in turn results in the enemy model falling behind the parent object meaning the model and hitbox get out of sync. This solution teleports the model back to local 0, 0, 0 after each loop of the walking animation which looks weird but is better than it being out of sync.