Enemy spawner (obelisk) can now be destroyed, fixed chest bug and improved dropped items
Enemy spawner now has health and can be killed just like an enemy, this includes working healthbar. Chest now work like they're supposed to. The inventory script no longer gives the player all items, player now only spawns with one axe in the hotbar. Dropped items will now use the model of the item that is selected in the inspector, the correct model is activated in the start method of the DroppedItem script.