Added simple quests, win screen, map
The quests just tell the player to find and retrieve three items. Once all three items have been picked up and put into the chest, the player wins. There are waypoints that show the direction of the items. The player can now, by pressing M, open a bigger map that shows the whole island. Dropped items now have an outline, both to show that they are special and to make them easier to see in the dark.